Compute CU for Sprinkler Irrigation Systems

Compute Christiansen Coefficient of Uniformity (CU or UC). Coefficient is based upon the average deviation from the mean value and the mean value.

CU(x)

Arguments

x

numeric array of catch can data. Array may include missing (NA) data. NA values will be removed in function.

Details

$$CU=100(1.0-\sum X/(nm))$$ where \(X=|z-m|\); absolute deviation of catch \(z\) from average catch \(m\) based on \(n\) observations (catches).

Value

CU value in percentage. Note that CU can take on a negative value if the average deviation from the mean is greater than the mean.

References

Christiansen, J. E. 1942. Irrigation by sprinkling. California Agricultural Experiment Station Bulletin 670. Berkeley, Cal.: University of California.

Keller, J., and R. D. Bliesner. 2000. Sprinkler and Trickle Irrigation. Caldwell, N.J.: Blackburn Press. ISBN 1-930665-19-9.

Author

Garry Grabow glgrabow@ncsu.edu

See also

Examples

# data below are volumes caught in ml.  Catch cans must be of
# equal surface area at top when using volume data
x<-c(47,42,45,24,13,26,33,34,27,30,40,44,32,12,12)
xcu<-CU(x)
#round results
curnd<-round(xcu,2)
xcu
#> [1] 69.13955
curnd
#> [1] 69.14