Compute distribution uniformity of lower half

Computes uniformity of lower half of a solid set sprinkler system vs. lower quarter as does DU. Note that this computation uses the same method as low quarter, except it uses the observations of the low half rather than low quarter. It does not use an empirical equation that converts DU (low quarter) to DUlh (lower half) as some have proposed.

Usage

DU.lh(x)

Arguments

x

a numeric array of catch depths or volumes. Array may include missing (NA) data. NA values will be removed in function.

Details

mean of lower half of ranked data divided by overall mean x 100.

Value

DU of lower half value in percentage.

Author

Garry Grabow glgrabow@ncsu.edu

See also

Examples

# data below are volumes caught in ml.  Catch cans must be of
# equal surface area at top when using volume data
x<-c(47,42,45,24,13,26,33,34,27,30,40,44,32,12,12)
xdulh<-DU.lh(x)
#round results
dulhrnd<-round(xdulh,2)
xdulh
#> [1] 71.58351
dulhrnd
#> [1] 71.58