Application efficiency and adequacy of a sprinkler irrigation system

determines amount of water above and below target depth to determine efficiency and adequacy from catch can data. If target depth is equal to soil moisture depletion in the root zone, all catch can depths greater than target are, in concept, lost to deep percolation and reduce efficiency, and all depths less than target are 100% efficient but reduce adequacy.

eff(x, target)

Arguments

x

numeric array of catch can depths

target

target depth

Details

computes efficiency and adequacy as amount of catch equal or less than target depth. Efficiency is defined and calculated as the applied (catch can caught) water retained in root zone divided by the total water applied when the target depth is set equal to the soil moisture deficit (SMD). Adequacy is the amount of catch equal or exceeding the target depth (adequacy). Amounts within the root zone and below the root zone are obtained using the results of `stats::density` using piecewise integration and normalized to target depth (normally SMD). Both efficiency and adequacy are reported in decimal (not percent) form.

Value

A named list with the following items:

appeff

application efficiency based on target depth, decimal

appadeq

application adequacy based on target depth, decimal

Author

Garry Grabow glgrabow@ncsu.edu

Note

If all catch can depths (and resultant density curve) are all below the target depth, the application efficiency is 100%. This assumes that the target depth is equal to the soil water deficit (or less). So deficit irrigation events are likely to be 100% or nearly 100% efficient.

See also

Examples

target<-3
# generate data with mean of target amount and high variability (non-uniformity)
x<-rnorm(25,3,1.75)
eff(x,target)
#> $appeff
#> [1] 0.6791853
#> 
#> $appadeq
#> [1] 0.8671671
#> 
# generate data with mean of target amount and low varibility (uniformity)
xx<-rnorm(25,3,0.5)
eff(xx,target)
#> $appeff
#> [1] 0.942701
#> 
#> $appadeq
#> [1] 0.9013608
#>