Application efficiency of the low quarter, (AELQ) for a sprinkler irrigation system

AELQ is a measure of both operation (duration of irrigation) and inherent sprinkler system uniformity. AELQ is based upon soil moisture depletion (SMD) at time of irrigation, the system application rate, and irrigation duration. If the low quarter caught depth is less than or equal to the SMD, AELQ is calculated as the low quarter average caught depth (from catch cans) divided by the average applied depth (sys. app. rate x duration). If the infiltrated (caught) low quarter depth is greater than SMD, AELQ is calculated as SMD/avg. applied depth in %. Whenever the irrigation (caught or infiltrated) in the low quarter exactly matches the SMD in the low quarter AELQ=PELQ. If the duration of irrigation is such that the infiltrated depth exceeds the SMD, then AELQ will be less than PELQ.

Usage

AELQ(x, rate,ss,sl,dur,smd,SI=TRUE)

Arguments

x

array of catch can caught rates - not depths (mm/hr or in/hr).

rate

sprinkler discharge rate (lpm or gpm).

ss

sprinkler spacing (m or ft).

sl

lateral spacing (m or ft).

dur

duration of irrigation event, hr

smd

soil moisture depletion at begin irrigation, mm or in.

SI

logical; units SI (mm, m, lpm) or US Customary (in., ft, gpm). SI (TRUE) is default.

x will be used to determine mode of AELQ computation by determining if the low quarter catch rate (and depth) is less than or greater than the SMD.

Value

AELQ (application efficiency of low quarter, %)

References

Mirriam and Keller, 1978. Farm System Irrigation Evaluation: A Guide for Management. PP 43,44. Utah State University, Logan, Utah. https://pdf.usaid.gov/pdf_docs/PNAAG745.pdf

Author

Garry Grabow glgrabow@ncsu.edu

See also

Examples

#see pp 41-44 of reference document
x<-c(.23,.31,.24,.24,.28,.31,.27,.22,.26,.31,
.31,.25,.20,.22,.32)# catch can depths caught during test converted to in/hr
SI<-FALSE # use U.S. customary units
smd<-4.4 #soil moisture depletion of 4.4 inches
rate<-4.6 # 4.6 gpm sprinkler discharge
sl<-50; ss<-30 # 30 x 50 ft sprinkler x lateral spacing
dur<-23.5 #23.5 hr duration (24 hour set)
AELQ(x,rate,ss,sl,dur,smd,SI)
#> [1] 63.40039
# now for a lower applied depth for alternate mode of AELQ
#computation
dur<-11.5 #change irrigation duration to a 12 hour set (11.5 hrs)
AELQ(x,rate,ss,sl,dur,smd,SI)
#> [1] 73.64892